Move get_reposters_for_playlist
